refactor : 알림 관련 엔티티 위치 조정
- notification 프로젝트 하위로 위치 조정
- BaseEntity, JpaAuditingConfig 추가
- Notification의 Member 연관관계를 삭제하고 memberId 키 참조로 변경
