+++ backend/notification/src/main/java/site/timecapsulearchive/notification/entity/Notification.java
@@ -1,24 +1,19 @@
-package site.timecapsulearchive.core.global.entity;
+package site.timecapsulearchive.notification.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import site.timecapsulearchive.core.domain.BaseEntity;
-import site.timecapsulearchive.core.domain.member.entity.Member;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "NOTIFICATIONS")
+@Table(name = "NOTIFICATION")
 public class Notification extends BaseEntity {
 
     @Id
@@ -29,8 +24,6 @@ public class Notification extends BaseEntity {
     @Column(nullable = false)
     private String text;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
+    @Column(name = "member_id")
+    private Long memberId;
 }
