+++ backend/notification/src/main/java/site/timecapsulearchive/notification/global/config/JpaAuditingConfig.java
@@ -1,0 +1,11 @@
+package site.timecapsulearchive.notification.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+
+}
+
