+++ backend/notification/src/main/java/site/timecapsulearchive/notification/repository/NotificationCategoryRepository.java
@@ -1,7 +1,7 @@
-package site.timecapsulearchive.core.global.repository;
+package site.timecapsulearchive.notification.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import site.timecapsulearchive.core.global.entity.NotificationCategory;
+import site.timecapsulearchive.notification.entity.NotificationCategory;
 
 public interface NotificationCategoryRepository extends JpaRepository<NotificationCategory, Long> {
