+++ backend/notification/src/main/java/site/timecapsulearchive/notification/repository/NotificationRepository.java
@@ -1,7 +1,7 @@
-package site.timecapsulearchive.core.global.repository;
+package site.timecapsulearchive.notification.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import site.timecapsulearchive.core.global.entity.Notification;
+import site.timecapsulearchive.notification.entity.Notification;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
